Pre-extend the stack when returning nodes
This could make a large difference with at least Perl 5.20, which seems
to encounter behaviour quadratic in the (large) number of returned nodes.
I don't have the dataset at hand to add a verification test.

In the rudimentary self-test, the new code has a tiny, tiny advantage
over the exiting code.

The diagnosis was made by my colleague P. Erlewein.